Format tables and edit buttons
Successfully created a function to format json data into a dict in the same format as the other tables we need to add and subtract. Revised the add and remove buttons to make use of that formatted json data. Some bugs still present, including that different variants of a nutrient aren't recognized as different.